one-page-app/flask/requirements.txt
flask>=2.2.2
gunicorn>=20.1.0
jinja2>=3.1.2
pyyaml>=6.0
pandas>=1.5.3
pandas-datareader>=0.10.0
redis>=4.4.2
pyrabbit>=1.1.0
celery==5.2.7
gevent==22.10.2
flower==1.2.0
